+++ test/generators.test.inc.js
@@ -27,7 +27,7 @@ describe('generator function', function() {
 		var fn = promisify(f);
 
         var promise = fn();
-		expect(promise).to.be.instanceof(promisify.Promise);
+		expect(promise).to.be.instanceof(Promise);
         return promise;
 	});